Refactor level formula configuration with modal handlers in LevelConfigHandler
